% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{author_extractor}
\alias{author_extractor}
\title{Extract author information from the XML
Note: extracts the information for _one_ author}
\usage{
author_extractor(paper, author_xml,
  authors_xpath = yaml::yaml.load_file(obkms$config$authors_db_xpath),
  document)
}
\arguments{
\item{paper}{URI of the research paper}

\item{author_xml}{part of the XML where the author information is found}
}
\description{
Extract author information from the XML
Note: extracts the information for _one_ author
}
