% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{bdj_dumper}
\alias{bdj_dumper}
\title{Dumps all of the BDJ up to the present date in a predefined directory}
\usage{
bdj_dumper()
}
\description{
Dumps all of the BDJ up to the present date in a predefined directory
}

