% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{begin_couplet}
\alias{begin_couplet}
\title{Probably junk}
\usage{
begin_couplet(subject, predicate, object)

begin_couplet(subject, predicate, object)
}
\description{
#' a function to start an rdf couplet

#' a function to start an rdf couplet
}

