% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{clear_context}
\alias{clear_context}
\title{Deletes a named graph named context}
\usage{
clear_context(context)
}
\arguments{
\item{context}{graph name}
}
\value{
status of query
}
\description{
Deletes a named graph named context
}

