% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{construct_nomen_triples}
\alias{construct_nomen_triples}
\title{Construct the nomen triples by creating a NomenclaturalAct}
\usage{
construct_nomen_triples(nomen_section_id, valid_name, nomen_citations)
}
\arguments{
\item{nomen_section_id}{the id of the nomen section to attach the act to}

\item{valid_name}{the single valid name}

\item{nomen_citations}{a vector of citations (synonyms, etc)}
}
\value{
triples
}
\description{
Construct the nomen triples by creating a NomenclaturalAct
}
