% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{construct_tnu_triples}
\alias{construct_tnu_triples}
\title{Connects one local entity, to a list of the names that have been used
via an intermediate class TNU}
\usage{
construct_tnu_triples(local_entity, names_used)
}
\arguments{
\item{local_entity}{the local entity which has used the names}

\item{names_used}{list of names that have been used}
}
\value{
triples
}
\description{
Connects one local entity, to a list of the names that have been used
via an intermediate class TNU
}

