% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{count_scientific_names}
\alias{count_scientific_names}
\title{Count how many distinct names there are}
\usage{
count_scientific_names()
}
\value{
number of names
}
\description{
Count how many distinct names there are
}
