% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{count_triples}
\alias{count_triples}
\title{How many statements do we have}
\usage{
count_triples(graph = "")
}
\arguments{
\item{graph}{from which graph}
}
\value{
number of triples in the graph
}
\description{
How many statements do we have
}
