% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{dbpedia_lookup}
\alias{dbpedia_lookup}
\title{Looks up the id of a concept in dbpedia}
\usage{
dbpedia_lookup(label, concept_type = "none", best_match = TRUE,
  lang = "English")
}
\description{
Looks up the id of a concept in dbpedia
}
