% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{delete_node_data}
\alias{delete_node_data}
\title{Delete all triples where a given node is a subject or object}
\usage{
delete_node_data(node_id)
}
\arguments{
\item{node_id}{id of the node to remove}
}
\value{
whatever the update call returns
TODO how to return the number of statements
}
\description{
Delete all triples where a given node is a subject or object
}
