% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{disambig_author}
\alias{disambig_author}
\title{Author Disambiguation}
\usage{
disambig_author(Literal, Identifier)
}
\arguments{
\item{Identifier}{a list of URI's associated with the author}
}
\value{
a URI of an author if it already exists, or mints new one if it
doesn't or there are multiple matches and thus an ambiguity.
}
\description{
This is a rule-based function. The function starts by looking at the most
effective rules and working itself to least effective rules.
}
\details{
Rule 1: If it is a collaborative author, the function does a direct match
of the author string.

Rule 2: For persons:
   2.a. fuzzy match of last name, and
   2.b. match beginning of first name,
   2.c. exact match of email.

Rule 3: match with institution - replace exact match of mailbox with exact match of institution URI

Rule 4: match with affiliation string

Args:
  @param Literal a list of literals associated with the author
}
