% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract.metadata}
\alias{extract.metadata}
\title{Extract metadata from an XML document but return value is a list}
\usage{
extract.metadata(xml, xlit)
}
\value{
\emph{list} of triples
}
\description{
Extract metadata from an XML document but return value is a list
}

