% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_front_matter}
\alias{extract_front_matter}
\title{Extract metadata from an XML2 object}
\usage{
extract_front_matter(xml, xlit)
}
\arguments{
\item{xml}{\emph{XML2} object}

\item{xlit}{\emph{list} of XPATH locations of literals entities}
}
\description{
Extract metadata from an XML2 object
}

