% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_metadata}
\alias{extract_metadata}
\title{Extract metadata from an XML2 object}
\usage{
extract_metadata(xml, x)
}
\arguments{
\item{xml}{\emph{XML2} object}

\item{x}{\emph{list} of XPATH locations of different entities}
}
\description{
Extract metadata from an XML2 object
}

