% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_nomen_citations}
\alias{extract_nomen_citations}
\title{Extract nomenclature citations (next citations) from a nomenclatural section in the XML
#' @param nomen_xml the nomenclatural node in the taxpub file}
\usage{
extract_nomen_citations(nomen_xml)
}
\value{
the OBKMS node of the valid (first name)
}
\description{
Extract nomenclature citations (next citations) from a nomenclatural section in the XML
#' @param nomen_xml the nomenclatural node in the taxpub file
}
