% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_publisher_info}
\alias{extract_publisher_info}
\title{A function to extract the publisher information from XML}
\usage{
extract_publisher_info(xml)
}
\description{
A function to extract the publisher information from XML
}

