% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_scientific_names_used}
\alias{extract_scientific_names_used}
\title{Extract taxonomic name usages from an XML nodeset
It also looks into descendant nodes thanks to the // in Xpath
Actually we want to submit nomenclature right away to OBKMS
so as not have duplicates.}
\usage{
extract_scientific_names_used(xml_ns)
}
\arguments{
\item{nodeset}{\emph{xml_nodeset} to check for TNU's}
}
\value{
\emph{list} of scientific name id's and triples with information
}
\description{
Extract taxonomic name usages from an XML nodeset
It also looks into descendant nodes thanks to the // in Xpath
Actually we want to submit nomenclature right away to OBKMS
so as not have duplicates.
}

