% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_single_scientific_name}
\alias{extract_single_scientific_name}
\title{Extracts only one scientific name from an XML node}
\usage{
extract_single_scientific_name(t)
}
\arguments{
\item{t}{the tp:taxon-name node}
}
\value{
the id of the name in OBKMS
TODO serialize this function as turtle
TODO after lookup, this function should be able to UPDATE the rank
of a name, if uncertain "insertae"
TODO also do authorship, etc.
}
\description{
Extracts only one scientific name from an XML node
}

