% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_tnu}
\alias{extract_tnu}
\title{Extract taxonomic name usages from an XML nodeset}
\usage{
extract_tnu(xml_ns)
}
\arguments{
\item{nodeset}{\emph{xml_nodeset} to check for TNU's}
}
\value{
\emph{list} of scientific name id's and triples with information
}
\description{
Extract taxonomic name usages from an XML nodeset
}

