% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_valid_name}
\alias{extract_valid_name}
\title{Extract the valid name (first citation) from a nomenclatural section in the XML}
\usage{
extract_valid_name(nomen_xml)
}
\arguments{
\item{nomen_xml}{the nomenclatural node in the taxpub file}
}
\value{
the OBKMS node of the valid (first name)
}
\description{
Extract the valid name (first citation) from a nomenclatural section in the XML
}
