% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{find_document_component_entities}
\alias{find_document_component_entities}
\title{Finds document components id and xml nodes}
\usage{
find_document_component_entities(xml, document_component_xpath)
}
\arguments{
\item{xml}{the xml resource}

\item{document_component_xpath}{datapase of xpaths where the document components
are found in the corresponding XML schema}
}
\description{
Finds document components id and xml nodes
}
