% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{find_local_xml}
\alias{find_local_xml}
\title{Finds the XML nodes corresponding to semantic entities}
\usage{
find_local_xml(xml, local_entities, local_entities_xpath)
}
\arguments{
\item{xml}{\emph{XML2} resource to look at}

\item{local_entities}{\emph{list} of semantic id's of entities}
}
\value{
\emph{list} of XML nodes corresponding to the entities
}
\description{
Finds the XML nodes corresponding to semantic entities
}

