% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{find_related_names}
\alias{find_related_names}
\title{Finds all related names to a given name
TODO : modify prepend_prefixes function to do both turtle and SPARQL}
\usage{
find_related_names(sci_name_label)
}
\arguments{
\item{sci_name_label}{the scientific name (label) for which to look for related names}
}
\value{
\emph{list} of related name-ids and their labels
}
\description{
Finds all related names to a given name
TODO : modify prepend_prefixes function to do both turtle and SPARQL
}
