% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_context_of}
\alias{get_context_of}
\title{Gets the graph name of an article of it exists}
\usage{
get_context_of(doi)
}
\arguments{
\item{doi}{the DOI of the article the context of which we are looking for}
}
\description{
Gets the graph name of an article of it exists
}

