% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{get_nodeid}
\alias{get_nodeid}
\title{Given a label, tries to get the GUID of the node that has that label}
\usage{
get_nodeid(label = "")
}
\arguments{
\item{options}{a list returned by \code{create_server_options}.}

\item{repo_id}{a}
}
\description{
We need a helper function to get or create an node (id) for the network given a preferred label (`skos:prefLabel`)
}
\examples{
\dontrun{}
}

