% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_institution_db}
\alias{init_institution_db}
\title{Initializes the educational institutions gazetteer from dbpedia}
\usage{
init_institution_db(fromFile = FALSE, offsetSize = 10000)
}
\value{
nothing
}
\description{
Args:
  @param fromFile if TRUE will read from a local file that has been given in
  the initial dump configuraiton
}
