% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{load_yaml_database}
\alias{load_yaml_database}
\title{Loads a package database from a yaml file
TODO needs review}
\usage{
load_yaml_database(database_path, internal_name)
}
\arguments{
\item{database_path}{path to the database}

\item{internal_name}{is the path to internal database to be used if none supplied}
}
\description{
Loads a package database from a yaml file
TODO needs review
}

