% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_event}
\alias{log_event}
\title{OBKMS Internal Logging Function}
\usage{
log_event(eventName, callingFunction, eventNotes = "",
  eventType = c("NOTE"), eventTime = Sys.time(),
  eventContext = obkms$log$current_context)
}
\value{
does not return anything, only sets the `log` list in the `OBKMS`
  environment
}
\description{
Args:
  @param message    the event message, i.e. what is going on
  @param ctime      current time, defaults to `Sys.time()`
  @param event_type type of event, one of
                      `c('NOTE', 'WARNING', 'ERROR', 'DEBUG')`
  @param context    the context in which the event is logged. For example it
                    could be the name of the function, or the name of the
                    workflow that is being run. Defaults to the current
                    context as found in `obkms$log$current_context`
}
