% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookup_author}
\alias{lookup_author}
\title{Looks up an author, given a list of literals}
\usage{
lookup_author(literals, concept_type = "none", best_match = TRUE,
  lang = "English")
}
\description{
Looks up an author, given a list of literals
}
