% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookup_id}
\alias{lookup_id}
\title{Lookup an Identifier in OBKMS}
\usage{
lookup_id(label, allow_multiple = FALSE, ...)
}
\arguments{
\item{label}{a vector of labels for which to look , the first one is preferred, all other ones are alternative;
doesn't have to specified. lookup can done only thru the ...}

\item{allow_multiple}{whether to allow to return multiple ids}

\item{...}{a list of additional things that can be used for look-up . read documentaiton carefully}
}
\description{
This function is similar to `get_nodeid`, but allows more flexibility.
Its most basic form is when the user specifies a `preferred_label`
and the system searches for all entities in the graph database that match
this preferred label exactly.
}
\examples{
\dontrun{}
}

