% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookup_institution}
\alias{lookup_institution}
\title{Lookup a URI of an institution given an affiliation string.}
\usage{
lookup_institution(affiliation, cluster)
}
\value{
URI of the institution if lookup was successful, NULL otherwise
}
\description{
Args:
  @param affiliation string to use to find the id of the organization
  @param cluster needed for the computation
}
