% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{majority_id}
\alias{majority_id}
\title{Most Frequent Id in a Dataframe}
\usage{
majority_id(aMatch, aFrame, idField = "id")
}
\value{
the URI (id) that matched the most rows or NULL if it is impossible
to calculate
}
\description{
Args:
  @param aMatch a numeric of rows matching in the data frame
  @param aFrame the data frame where the matches are sought
  @param idField what is the name of the column where the id's are stored
}
