% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{parse_affiliation}
\alias{parse_affiliation}
\title{Parses an affiliation string}
\usage{
parse_affiliation(current_affiliation)
}
\arguments{
\item{current_affiliation}{unparsed affiliation string as found in the XML}
}
\value{
a list of three - institution, city, and country
}
\description{
Parses an affiliation string
}
