% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{prefix_serializer}
\alias{prefix_serializer}
\title{Serializes the prefix database in a language}
\usage{
prefix_serializer(reqd_prefixes, prefix_db = paste0(path.package("obkms"),
  "/prefix_db.yml"), prefix_lang = "Turtle")
}
\arguments{
\item{reqd_prefixes}{a character vector of needed prefixes, can be missing}

\item{prefix_db}{path to a YAML prefix database, has a default from the pkg}

\item{prefix_lang}{whether you want SPARQL or Turtle, the default is SPARQL}
}
\description{
#' If `reqd_prefixes` is left missing, all prefixes will be returned.
The individual prefixes may or may not end with ":". If they don't,
it will be added during function execution.
}

