% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{prefix_ttl}
\alias{prefix_ttl}
\title{Returns a prefix string in Turtle}
\usage{
prefix_ttl(reqd_prefixes, prefix_db = paste0(path.package("obkms"),
  "/prefix_db.yml"))
}
\arguments{
\item{reqd_prefixes}{a list of needed prefixes}
}
\description{
Returns a prefix string in Turtle
}

