% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{prepend_prefixes}
\alias{prepend_prefixes}
\title{Use the prefix database to create Turtle statements}
\usage{
prepend_prefixes(t = c("turtle"))
}
\arguments{
\item{t}{the syntax}
}
\description{
Use the prefix database to create Turtle statements
}

