% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{qname}
\alias{qname}
\title{''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
Minimizes a URI to a Qname.}
\usage{
qname(uri)

qname(uri)
}
\arguments{
\item{uri}{the full URI to minimize}

\item{uri}{the full URI to minimize}
}
\description{
''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''''
Minimizes a URI to a Qname.

Minimizes a URI to a Qname.
}

