% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{strip_angle}
\alias{strip_angle}
\title{Strips Angular Brackets from a URI if Present}
\usage{
strip_angle(uri, reverse = FALSE)
}
\arguments{
\item{uri}{the URI to strip angular brackets from}

\item{reverse}{FALSE, if true will put angular brakcets instead}
}
\value{
URI with stripped angular brackets around it
TODO this function should not be exported
}
\description{
Strips Angular Brackets from a URI if Present
}
