% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{submit_turtle}
\alias{submit_turtle}
\title{Submits RDF string (in Turtle syntax) to OBKMS}
\usage{
submit_turtle(rdf)
}
\arguments{
\item{rdf}{the RDF string}
}
\description{
Submits RDF string (in Turtle syntax) to OBKMS
}

