% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{taxpub_extractor}
\alias{taxpub_extractor}
\title{Top Level Function to Convert TaxPub to Triples}
\usage{
taxpub_extractor(xml, xlit, xdoco)
}
\arguments{
\item{xml}{an XML2 document}

\item{xlit}{a of XPATH locations of literals entities (atoms) of a TaxPub document}

\item{xdoco}{list of XPATH locations of connected entities (not atoms) that ought to be processed via lower level extractors}
}
\value{
list of named triples lists, the name corresponds to the context TODO check what's actually retuerned
}
\description{
Top Level Function to Convert TaxPub to Triples
}

