% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{trim_label}
\alias{trim_label}
\title{Trims a label of extraneous white space characters}
\usage{
trim_label(label)
}
\arguments{
\item{label}{what to trim}
}
\value{
the "label" with traling or leading whitespace removed, and any
extra white space within is removed
}
\description{
Trims a label of extraneous white space characters
}
