% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{triple}
\alias{triple}
\title{Ensure correctness of a triple}
\usage{
triple(S, P, O, blank = FALSE)
}
\arguments{
\item{S}{\emph{character} subject}

\item{P}{\emph{character} predicate}

\item{O}{\emph{character OR list} object}
}
\value{
NULL, if one of the values is missing, a character of length 3 otherwise
unless blank is TRUE, then S = ""
}
\description{
Ensure correctness of a triple
}

