% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{update_authors_db}
\alias{update_authors_db}
\title{Updates Local Authors Store}
\usage{
update_authors_db(aId, aAuthor, isPerson, firstName, familyName, eMail, aInst,
  aAff)
}
\arguments{
\item{aId}{author URI}

\item{aAuthor}{author label}

\item{isPerson}{is it a person or an agent}

\item{aInst}{author's institution URI}

\item{aAff}{author's affiliation string}
}
\description{
Updates Local Authors Store
}
