% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{write_couplet}
\alias{write_couplet}
\title{Writes a single couplet for the given subejct}
\usage{
write_couplet(subject, triples)

write_couplet(subject, triples)
}
\description{
Writes a single couplet for the given subejct

Writes a single couplet for the given subejct
}

