% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{write_couplet2}
\alias{write_couplet2}
\title{Writes a single couplet for the given subejct
with triples in a list}
\usage{
write_couplet2(subject, triples)

write_couplet2(subject, triples)
}
\description{
Writes a single couplet for the given subejct
with triples in a list

Writes a single couplet for the given subejct
with triples in a list
}

