% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{write_log}
\alias{write_log}
\title{Write an OBKMS Iteration Log}
\usage{
write_log(toFile = FALSE, append = TRUE, iteration_name = "")
}
\arguments{
\item{toFile}{\emph{logical}: if this parameter is TRUE, then the log is
written in a text-file}

\item{append}{\emph{logical}: if this parameter is TRUE the log will be
appended to the file}

\item{iteration_name}{this string will be last part of the filename}
}
\value{
a data.frame with the log
}
\description{
This function both writes the log to a file and returns it as a data.frame
}
