% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{write_predicate_stanza}
\alias{write_predicate_stanza}
\title{Writes the predicate-object stanza after a subject has been written}
\usage{
write_predicate_stanza(predicate, triples)
}
\description{
Writes the predicate-object stanza after a subject has been written
}

