% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{write_predicate_stanza2}
\alias{write_predicate_stanza2}
\alias{write_predicate_stanza2}
\title{takes care of the predicates, but with the triple structure a list}
\usage{
write_predicate_stanza2(predicate, triples)

write_predicate_stanza2(predicate, triples)
}
\description{
takes care of the predicates, but with the triple structure a list

takes care of the predicates, but with the triple structure a list
}
