% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{xml2rdf}
\alias{xml2rdf}
\title{Convert an XML document to RDF}
\usage{
xml2rdf(resource_locator, resource_type = "FILE",
  resource_format = "TAXPUB", serialization_format = "TURTLE")
}
\arguments{
\item{resource_locator}{address/ location of the XML resource}

\item{resource_format}{XML schema of the XML resource;
one of { "TAXPUB", "REFBANK_XML" }}

\item{serialization_format}{output serialization format of the RDF;
one of "TURTLE"}

\item{resouce_type}{type of location of the XML resource;
one of "FILE"}
}
\description{
Convert an XML document to RDF
}
\author{
Viktor Senderov
}

