% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{dbpedia_lookup}
\alias{dbpedia_lookup}
\title{Looks up the URI of a Resource in DBPedia}
\usage{
dbpedia_lookup(label, label_property = obkms$properties$label,
  resource_type = obkms$classes$Thing,
  language = obkms$parameters$Language$English, print_query = FALSE)
}
\arguments{
\item{label}{the label of the resource}

\item{label_property}{the type of label (i.e. rdfs:label, skos:prefLabel,
skos:altLabel that we are looking for) needs to have a `uri` fields
as in obkms$properties. Default is `rdfs:label`}

\item{resource_type}{the type of the resource as specified by OBKMS classes,
needs to be a list with a named field called URI. Default is Thing.}

\item{language}{of the label as per obkms$parameters$Language Needs to have a
`semantic_code` field. Default is English.}

\item{print_query}{prints out the query to the console for debugging purposes
Default is FALSE.}
}
\value{
one or more matches of the search criteria, NULL if nothing is matched
}
\description{
Looks up the URI of a Resource in DBPedia
}
\examples{

dbpedia_lookup("Abraham Lincoln")

BiologicalThing = obkms$classes$DbpediaBiologicalThing
dbpedia_lookup("Panthera", resource_type = BiologicalThing )

}
