% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{extract_information}
\alias{extract_information}
\title{Top level extractor}
\usage{
extract_information(xml, xlit, xdoco)
}
\arguments{
\item{xml}{\emph{XML2} object}

\item{xlit}{\emph{list} of XPATH locations of literals entities}
}
\value{
\emph{list} of named triples lists, the name corresponds to the
context
}
\description{
Top level extractor
}
