% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{init_env}
\alias{init_env}
\title{Initialize package parameters upon loading.}
\usage{
init_env(server_access_options, prefix_db = paste0(path.package("obkms"), "/",
  "prefix_db.yml"), entities_db = paste0(path.package("obkms"), "/",
  "semantic_entities_db.yml"),
  literals_db_xpath = paste0(path.package("obkms"), "/",
  "literals_db_xpath.yml"),
  non_literals_db_xpath = paste0(path.package("obkms"), "/",
  "non_literals_db_xpath.yml"), xml_source = "file", xml_type = "taxpub")
}
\description{
The package contains a database of semantic web prefixes, stored in a yaml
database in the `inst/` folder. The name of the database file is assigned to
the package environment when the package is loaded.
Store graph database access data in an environment.
}
\details{
In order not to have to parameterize each function with the database access
information, the access options are stored in an environment, which is then
utilzed by the individual package functions. It also enriches the options
}

