% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{lookup_id}
\alias{lookup_id}
\title{Lookup an Identifier (URI) for a Resource in OBKMS}
\usage{
lookup_id(label, article_id = "?context", language = NULL,
  resource_type = obkms$classes$Thing, ignore_case = FALSE,
  generate_on_fail = TRUE, trim = TRUE, show_query = FALSE)
}
\value{
matching ID's or NULL
}
\description{
Given a label (and a few optional arguments) this function locates the URI
or URI's that match this label (and optional arguments) in OBKMS. If no
match is found, a unique ID (UUID) is generated.
}
\details{
Args:
  @param label label for which to look
  @param article_id what article is this label related to
    (used as a context in the graph patttern), if no article id is supplied,
    will look in the default graph
  @param language default is English
  @param ignore_case if TRUE lowercases everything
  @param generate_on_fail if TRUE if no match is found, a new ID is generated,
    if FALSE, NULL is returned
  @param trim if TRUE whitespace will be trimmed inside and before and after
    the label
}
\examples{

lookup_id( "10.3897/BDJ.1.e1000" )

}
