% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdf.R
\name{qname}
\alias{qname}
\title{Minimizes a URI to a Qname.}
\usage{
qname(uri)
}
\arguments{
\item{uri}{the full URI to minimize}
}
\description{
E.g. qname("http://openbiodiv.net/leis-papuensis") minimizes to
":leis-papuensis". The special case here is the `_base` prefixes, as
it minimizes to nothing.
}
\details{
if the prefix is _base, it is shortened to :
}

