% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xml.R
\name{taxpub_extractor}
\alias{taxpub_extractor}
\title{Top Level Extractor from TaxPub to Triples}
\usage{
taxpub_extractor(xml,
  xlit = yaml::yaml.load_file(obkms$config$literals_db_xpath),
  xdoco = yaml::yaml.load_file(obkms$config$non_literals_db_xpath))
}
\arguments{
\item{xml}{an XML document (as parsed by the `xml2` package)}

\item{xlit}{XPATH locations of atoms (literals) of a TaxPub document}

\item{xdoco}{XPATH locations of document entities, parts of the document that are not atoms and are identifiers}
}
\value{
TODO write what is returned
}
\description{
Top Level Extractor from TaxPub to Triples
}
