% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{write_log}
\alias{write_log}
\title{Write the OBKMS to log to a file}
\usage{
write_log(toFile = FALSE)
}
\value{
the dataframe of the log
}
\description{
Args:
  @param toFile if this parameter is TRUE, then the log is also appended
                to a text file
}
